CiudadesPage
Arreglado diseño de la vista al no encontrar estado

+++ src/pages/DashboardCiudadesPage.js
@@ -1,19 +1,18 @@
-import { map } from "@firebase/util";
 import React from "react";
-import CiudadesLista from "../components/CiudadesLista/CiudadesLista";
 import '../css/dashboardCiudades.css'
-import {queryCiudades, deleteCiudad, createCiudad} from "../services/ciudades"
+import { queryCiudades, deleteCiudad, createCiudad } from "../services/ciudades"
 import { useNavigate } from 'react-router-dom';
+import swal from "sweetalert";
+import { CircularProgress } from "@material-ui/core"
 
 
-function DashboardCiudadesPage(){
+function DashboardCiudadesPage() {
 
     const [ciudades, setCiudades] = React.useState([])
     const [loading, setLoading] = React.useState(false)
     const navigate = useNavigate()
 
     React.useEffect(() => {
-
         setLoading(true)
         queryCiudades().then((response) => {
             setCiudades(response)
@@ -22,14 +21,33 @@ function DashboardCiudadesPage(){
 
     }, [])
 
-    
 
-    const handleDelete = (ciudadId) =>{
-        setLoading(true)
-        deleteCiudad(ciudadId).then(() => {
-            setCiudades(ciudades.filter( ciudad => ciudadId != ciudad.id))
-            setLoading(false)
+
+    const handleDelete = (ciudadId) => {
+        swal({
+            title: "¿Estás seguro?",
+            text: "Una vez eliminado, no se podrá recuperar los datos almacenados",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
         })
+            .then((willDelete) => {
+                if (willDelete) {
+                    swal("Se eliminó con exito.", {
+                        icon: "success",
+
+                    }
+
+                    );
+                    setLoading(true)
+                    deleteCiudad(ciudadId).then(() => {
+                        setCiudades(ciudades.filter(ciudad => ciudadId != ciudad.id))
+                        setLoading(false)
+                    })
+                } else {
+                    swal("Your imaginary file is safe!");
+                }
+            });
     }
 
 
@@ -45,59 +63,59 @@ function DashboardCiudadesPage(){
     }
 
     if (loading) return (
-        <h3>Cargando</h3>
+        <CircularProgress/>
     )
 
     return (
         <>
-        <div className="imagenFondo">
-            <div className="cont">
-                <div className="contenedor">
-                    <table className="tabla">
-                        <tr className="head">
+            <div className="imagenFondo">
+                <div className="cont">
+                    <div className="contenedor">
+                        <table className="tabla">
+                            <tr className="head">
 
-                            <th >Nombre</th>
+                                <th >Ciudad</th>
 
-                            <th>Accion</th>
+                                <th>Accion</th>
 
-        
 
-                        </tr>
-                        {loading ? <div>cargando</div> : null}
-                        {ciudades.map(({id,nombre})=>(
 
-                            <>
-                                <tr border>
-                                    <td className="Nombre">
-                                        {nombre}
-                                    </td>
-                                    <td>
-                                        <div className="botonesContainer">
-                                            <button className='button' onClick={() => navigate(`/dashboardCiudades/${id}`)}> Editar </button>
-                                            <button className='buttonE' onClick={(() => {handleDelete(id)})}> Eliminar </button>
-                                        </div>
-                                    </td>
-                                </tr>
+                            </tr>
+                            {loading ? <CircularProgress/> : null}
+                            {ciudades.map(({ id, nombre }) => (
+
+                                <>
+                                    <tr border>
+                                        <td className="Nombre">
+                                            {nombre}
+                                        </td>
+                                        <td>
+                                            <div className="botonesContainer">
+                                                <button className='button' onClick={() => navigate(`/dashboardCiudades/${id}`)}> Editar </button>
+                                                <button className='buttonE' onClick={(() => { handleDelete(id) })}> Eliminar </button>
+                                            </div>
+                                        </td>
+                                    </tr>
+
+                                </>
+                            ))}
+                        </table>
+
+                        <div>
+
                             
-                            </>
-                        ))}
-                    </table>
+                            <form onSubmit={handleSubmit}>
 
-                <div>
-                    
-                    <div className="title">Agregar una nueva ciudad </div>
-                    <form onSubmit={handleSubmit}>
-                        
-                        <button className='button'onClick = {() => navigate(`/dashboardCiudades/create`)}>Submit</button>
-                    </form> 
-                    <br/>
-                    <div>Volver al Dashboard</div>
-                    <button className='button' onClick={() => navigate(`/dashboard`)}>Click!!!</button>
+                                <button className='button' onClick={() => navigate(`/dashboardCiudades/create`)}>Agregar ciudad</button>
+                            </form>
+                            <br />
+                            
+                            <button className='button' onClick={() => navigate(`/dashboard`)}>Volver al dashboard</button>
+                        </div>
                     </div>
                 </div>
-            </div>    
-        </div>
-    </>
+            </div>
+        </>
     );
 
 }
